requirements.txt
streamlit
PyMuPDF
pytesseract
pdf2image
Pillow
google-generativeai
openpyxl
